ocr/requirements.txt
torch>=2.0
torchvision>=0.15
opencv-python-headless>=4.8
Pillow>=10.0
numpy>=1.24
